requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
openai==2.3.0
packaging==25.0
pydantic==2.12.0
pydantic_core==2.41.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0+cpu
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
